+++ R/htmlTable_render_addCells.R
@@ -17,10 +17,12 @@
 #' @return `string` Returns the string with the new cell elements
 #' @keywords internal
 #' @family hidden helper functions for htmlTable
-prAddCells <- function(rowcells, cellcode, style_list, style, prepped_cell_css, cgroup_spacer_cells, has_rn_col, offset = 1) {
+#' @importFrom stringr str_interp
+prAddCells <- function(rowcells, cellcode, style_list, style, prepped_cell_css, cgroup_spacer_cells, has_rn_col, offset = 1, style_list_align_key = "align") {
   cell_str <- ""
   style <- prAddSemicolon2StrEnd(style)
 
+  previous_was_spacer_cell <- FALSE
   for (nr in offset:length(rowcells)) {
     cell_value <- rowcells[nr]
     # We don't want missing to be NA in a table, it should be empty
@@ -28,35 +30,45 @@ prAddCells <- function(rowcells, cellcode, style_list, style, prepped_cell_css, 
       cell_value <- ""
     }
 
-    cell_style <- c(
-      prepped_cell_css[nr],
-      style,
-      prGetAlign(style_list$align, index = nr + has_rn_col)
-    )
+    followed_by_spacer_cell <- nr != length(rowcells) &&
+      nr <= length(cgroup_spacer_cells) &&
+      cgroup_spacer_cells[nr] > 0
+
+    align_style <- prGetAlign(style_list[[style_list_align_key]],
+                              index = nr + has_rn_col,
+                              style_list = style_list,
+                              followed_by_spacer_cell = followed_by_spacer_cell,
+                              previous_was_spacer_cell = previous_was_spacer_cell)
+    cell_style <- c(prepped_cell_css[nr],
+                    style)
+
     if (!is.null(style_list$col.columns)) {
       cell_style %<>%
         c(`background-color` = style_list$col.columns[nr])
     }
 
-    cell_str %<>%
-      sprintf(
-        "%s\n\t\t<%s style='%s'>%s</%s>",
-        .,
-        cellcode,
-        prGetStyle(cell_style),
-        cell_value,
-        cellcode
-      )
+
+    cell_str %<>% paste(str_interp("<${CELL_TAG} style='${STYLE}'>${CONTENT}</${CELL_TAG}>",
+                                   list(CELL_TAG = cellcode,
+                                        STYLE = prGetStyle(cell_style,
+                                                           align_style),
+                                        CONTENT = cell_value)),
+                        sep = "\n\t\t")
 
     # Add empty cell if not last column
-    if (nr != length(rowcells) &&
-      nr <= length(cgroup_spacer_cells) &&
-      cgroup_spacer_cells[nr] > 0) {
+    if (followed_by_spacer_cell) {
+      align_style <- prGetAlign(style_list[[style_list_align_key]],
+                                index = nr + has_rn_col,
+                                style_list = style_list,
+                                spacerCell = TRUE,
+                                followed_by_spacer_cell = followed_by_spacer_cell,
+                                previous_was_spacer_cell = previous_was_spacer_cell)
+
       # The same style as previous but without align borders
       cell_style <- c(
         prepped_cell_css[nr],
         style,
-        prGetAlign(style_list$align, index = nr + has_rn_col, emptyCell = TRUE)
+        align_style
       )
       spanner_style <- style
 
@@ -67,15 +79,15 @@ prAddCells <- function(rowcells, cellcode, style_list, style, prepped_cell_css, 
       }
 
       cell_str %<>%
-        sprintf(
-          "%s\n\t\t<%s style='%s' colspan='%d'>&nbsp;</%s>",
-          .,
-          cellcode,
-          prGetStyle(cell_style, spanner_style),
-          cgroup_spacer_cells[nr],
-          cellcode
-        )
+        paste("\n\t\t") %>%
+        prAddEmptySpacerCell(style_list = style_list,
+                             cell_style = prGetStyle(cell_style, spanner_style),
+                             colspan = cgroup_spacer_cells[nr],
+                             cell_tag = cellcode,
+                             align_style = align_style)
     }
+
+    previous_was_spacer_cell <- followed_by_spacer_cell
   }
   return(cell_str)
 }