+++ R/htmlTable_style_handlers.R
@@ -54,7 +54,7 @@
 #' @param css.header.border_bottom The header bottom-border style, e.g. `border-bottom: 1px solid grey`
 #' @param css.cgroup The same as `css.class` but for cgroup formatting.
 #' @param css.class The html CSS class for the table. This allows directing html
-#'  formatting through [CSS](http://www.w3schools.com/Css/)
+#'  formatting through [CSS](https://www.w3schools.com/Css/)
 #'  directly at all instances of that class. *Note:* unfortunately the
 #'  CSS is frequently ignored by word processors. This option
 #'  is mostly inteded for web-presentations.
@@ -73,6 +73,21 @@
 #' @param padding.tspanner The table spanner is usually without padding but you may specify padding
 #'  similar to `padding.rgroup` and it will be added to all elements, including the rgroup elements.
 #'  This allows for a 3-level hierarchy if needed.
+#' @param spacer.celltype When using cgroup the table headers are separated through a empty
+#'  HTML cell that is by default filled with `&nbsp;` (no-breaking-space) that prevents the cell
+#'  from collapsing. The purpose of this is to prevent the headers underline to bleed into one
+#'  as the underline is for the entire cell. You can alter this behavior by changing this option,
+#'  valid options are `single_empty`, `skip`, `double_cell`. The `single_empty` is the default,
+#'  the `skip` lets the header bleed into one and skips entirely, `double_cell` is for having
+#'  two cells so that a vertical border ends up centered (specified using the `align` option).
+#'  The arguments are matched internally using [base::match.arg] so you can specify only a part
+#'  of the name, e.g. `"sk"` will match `"skip"`.
+#' @param spacer.css.cgroup.bottom.border Defaults to `none` and used for separating cgroup headers.
+#'  Due to a browser bug this is sometimes ignored and you may therefore need to set this
+#'  to `1px solid white` to enforce a white border.
+#' @param spacer.css If you want the spacer cells to share settings you can set it here
+#' @param spacer.content Defaults to `&nbsp;` as this guarantees that the cell is not collapsed
+#'  and is highly compatible when copy-pasting to word processors.
 #'
 #' @return `x` with the style added as an attribute that the htmlTable then can use for formatting.
 #' @export
@@ -115,7 +130,11 @@ addHtmlTableStyle <- function(x,
 
                               # More alternatives
                               padding.rgroup = NULL,
-                              padding.tspanner = NULL) {
+                              padding.tspanner = NULL,
+                              spacer.celltype = NULL,
+                              spacer.css.cgroup.bottom.border = NULL,
+                              spacer.css = NULL,
+                              spacer.content = NULL) {
   style_list <- prGetAttrWithDefault(x,
     which = style_attribute_name,
     default = getHtmlTableTheme()
@@ -163,7 +182,11 @@ appendHtmlTableStyle <- function(x,
 
                                  # More alternatives
                                  padding.rgroup = NULL,
-                                 padding.tspanner = NULL) {
+                                 padding.tspanner = NULL,
+                                 spacer.celltype = NULL,
+                                 spacer.css.cgroup.bottom.border = NULL,
+                                 spacer.css = NULL,
+                                 spacer.content = NULL) {
   style_list <- prGetAttrWithDefault(x,
     which = style_attribute_name,
     default = getHtmlTableTheme()
@@ -179,6 +202,26 @@ appendHtmlTableStyle <- function(x,
   return(x)
 }
 
+#' Get style options for object
+#'
+#' A wrap around the [base::attr()] that retrieves the style
+#' attribute used by [htmlTable()] (`htmlTable.style`).
+#'
+#' @param x The object intended for [htmlTable()].
+#' @return A `list` if the attribute exists, otherwise `NULL`
+#' @export
+#' @examples
+#' library(magrittr)
+#'
+#' mx <- matrix(1:4, ncol = 2)
+#' colnames(mx) <- LETTERS[1:2]
+#' mx %>%
+#'   addHtmlTableStyle(align = "l|r") %>%
+#'   getHtmlTableStyle()
+getHtmlTableStyle <- function(x) {
+  attr(x, style_attribute_name, exact = TRUE)
+}
+
 #' Check if object has a style set to it
 #'
 #' If the attribute `htmlTable.style` is set it will check if
@@ -188,8 +231,17 @@ appendHtmlTableStyle <- function(x,
 #' @param style_name A string that contains the style name.
 #' @return `logical` `TRUE` if the attribute and style is not `NULL`
 #' @export
+#' @family htmlTableStyle
+#' @examples
+#' library(magrittr)
+#'
+#' mx <- matrix(1:4, ncol = 2)
+#' colnames(mx) <- LETTERS[1:2]
+#' mx %>%
+#'   addHtmlTableStyle(align = "l|r") %>%
+#'   hasHtmlTableStyle("align")
 hasHtmlTableStyle <- function(x, style_name) {
-  style <- attr(x, style_attribute_name, exact = TRUE)
+  style <- getHtmlTableStyle(x)
   if (is.null(style)) {
     return(FALSE)
   }
@@ -203,6 +255,7 @@ hasHtmlTableStyle <- function(x, style_name) {
 
 style_attribute_name <- "htmlTable.style"
 
+#' @importFrom stringr str_replace
 prValidateAndMergeStyles <- function(org_style_list, styles_from_arguments, overwrite) {
   assert_list(org_style_list)
   assert_list(styles_from_arguments)
@@ -213,19 +266,24 @@ prValidateAndMergeStyles <- function(org_style_list, styles_from_arguments, over
     return(org_style_list)
   }
 
+  default_args <- list("pos.caption" = c("top", "bottom", "below"),
+                       "spacer.celltype" =  c("single_empty", "skip", "double_cell"))
+
   style_list <- org_style_list
   for (n in names(styles_from_arguments)) {
     # We only merge css components when we're not replacing everything
     if (startsWith(n, "css") && n != "css.class" && !overwrite) {
       # The second argument takes precedence as the final style when conflicts arise
       style_list[[n]] <- prGetStyle(style_list[[n]], styles_from_arguments[[n]])
-    } else if (n == "pos.caption") {
-      valid_caption_positions <- c("top", "bottom", "below")
-      if (styles_from_arguments[[n]] %in% valid_caption_positions) {
-        style_list[[n]] <- styles_from_arguments[[n]]
-      } else {
-        stop("The argument ", n, " must be one of: ", paste(valid_caption_positions, collapse = ", "))
-      }
+    } else if (n %in% names(default_args)) {
+      tryCatch({
+        style_list[[n]] <- match.arg(arg = styles_from_arguments[[n]], choices = default_args[[n]])
+      }, error = function(x) {
+        x %>%
+          extract2("message") %>%
+          str_replace("^'arg'", sprintf("'%s'", argument_name = n)) %>%
+          stop()
+      })
     } else {
       style_list[[n]] <- styles_from_arguments[[n]]
     }
@@ -233,6 +291,7 @@ prValidateAndMergeStyles <- function(org_style_list, styles_from_arguments, over
 
   return(style_list)
 }
+
 
 prGetAttrWithDefault <- function(x, which, default = NA) {
   if (which %in% names(attributes(x))) {